Bones anim not working yet

+++ engine/core/include/models/bone.h
@@ -47,7 +47,7 @@ namespace engine
 			for (int positionIndex = 0; positionIndex < m_NumPositions; ++positionIndex)
 			{
 				aiVector3D aiPosition = channel->mPositionKeys[positionIndex].mValue;
-				float timeStamp = channel->mPositionKeys[positionIndex].mTime;
+				double timeStamp = channel->mPositionKeys[positionIndex].mTime;
 				KeyPosition data;
 				data.position = AssimpGLMHelpers::GetGLMVec(aiPosition);
 				data.timeStamp = timeStamp;
@@ -58,7 +58,7 @@ namespace engine
 			for (int rotationIndex = 0; rotationIndex < m_NumRotations; ++rotationIndex)
 			{
 				aiQuaternion aiOrientation = channel->mRotationKeys[rotationIndex].mValue;
-				float timeStamp = channel->mRotationKeys[rotationIndex].mTime;
+				double timeStamp = channel->mRotationKeys[rotationIndex].mTime;
 				KeyRotation data;
 				data.orientation = AssimpGLMHelpers::GetGLMQuat(aiOrientation);
 				data.timeStamp = timeStamp;
@@ -69,7 +69,7 @@ namespace engine
 			for (int keyIndex = 0; keyIndex < m_NumScalings; ++keyIndex)
 			{
 				aiVector3D scale = channel->mScalingKeys[keyIndex].mValue;
-				float timeStamp = channel->mScalingKeys[keyIndex].mTime;
+				double timeStamp = channel->mScalingKeys[keyIndex].mTime;
 				KeyScale data;
 				data.scale = AssimpGLMHelpers::GetGLMVec(scale);
 				data.timeStamp = timeStamp;
